pkg: bsp-imx: Add support to cross compilation
bsp-imx package builds u-boot and ATF for different i.MX8M platforms, which
can be very time consuming. This commit adds support to cross compilation
so it can build faster on x86. This package is only useful for ARM64, so a
check is added to skip calling make for any other architecture.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>