Introduce tee-utils package
The tee-utils package builds all TEE artifacts for userspace:
tee-supplicant, libtee libraries and xtest test suit.

A single tarball with all components is exported for each platform. This
tarball can be extracted in the rootfs of a EVE's debug container (for
instance) in order to execute and test trusted applications.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>