pillar: Improve parsing of environment variables
This commit improves the parsing of environment variables for Edge
Applications received from the controller. The go-envparse library is used
to detect syntax errors that would break the shell of the shim VM (in case
of containers) and to improve error messages. The checks are done in the
domain manager level.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>