tools: Introduces get-deps.sh script
This commit introduces the script get-deps.sh, which scans the pkg
directory and process each Dockerfile (or Dockerfile.in) file in order to
find all build dependencies (containers) of each package. It can also check
package dependencies for rootfs images. The final package dependency tree
can be generated as a PNG image (using dot tool) or as a Makefile that can
be included to the main Makefile.

The usage of the script is the following:

    ./tools/get-deps.sh [-r] <-i|-m> <output_file>

Options:
        -i <output_file> : Generate a PNG image file
        -m <output_file> : Generate a Makefile auxiliary file
        -r               : Also generates dependencies for rootfs images

Examples:

    ./tools/get-deps.sh -i deptree.png
    ./tools/get-deps.sh -r -i deptree.png
    ./tools/get-deps.sh -m pkg-deps.mk

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>