Makefile: Scan package dependency over pkg dir.
Container packages might import other containers (build/run dependencies,
etc) using FROM command inside their Dockerfile. However, these
dependencies are not exposed to the main Makefile. Using the script
tools/get-deps.sh the dependency graph can be generated into a file that
can be included in the main Makefile in order to take these package
dependencies into account. The script is also able to scan rootfs image
container dependencies.

This commit changes the main Makefile to call tools/get-deps.sh script in
order to generate a file called pkg-deps.mk with the package dependency
included, so we can ensure all dependencies from a package will be built
(when changed) in the correct order to avoid build failures, as can be
experienced specially when doing batch updates of the eve-alpine container
to multiple packages.

Unfortunately, rootfs images usually have a lot of package dependencies,
leading to a lot of docker build calls when all dependencies are checked,
which can increase substantially the build time. In order to void that, the
main Makefile will scan rootfs image dependencies only when the
ROOTFS_DEPS=y is specified, for example:

    make ROOTFS_DEPS=y rootfs

    make ROOTFS_DEPS=y pkgs

In summary:

    - ROOTFS_DEPS not defined: Only package dependencies are scanned. For
      example, if pkg/pillar depends on pkg/uefi and pkg/dnsmasq to be
      built and one or both of these packages are changed, then when doing
      make pkg/pillar, make will ensure the dependencies will be built
      before pkg/pillar.

    - ROOTFS_DEPS defined: The dependencies of all packages and the
      rootfs images will be scanned, which means that any dependency
      package that was changed will be built before the main target (pkgs,
      rootfs, live, etc).

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>