tools: Introduces get-deps tool
This commit introduces the tool get-deps (under tools/get-deps/), which
scans the pkg directory and process each Dockerfile (or Dockerfile.in) file
in order to find all build dependencies (containers) of each package. It
can also check package dependencies for rootfs images. The final package
dependency tree can be generated as a PNG image (using dot tool) or as a
Makefile that can be included to the main Makefile.

The usage of the tool the following:

    ./tools/get-deps/get-deps [-r] <-i|-m> <output_file>

  -i string
    	Generate a PNG image file
  -m string
    	Generate a Makefile auxiliary file
  -r	Also generates dependencies for rootfs image

Examples:

    ./tools/get-deps/get-deps -i deptree.png
    ./tools/get-deps/get-deps -r -i deptree.png
    ./tools/get-deps/get-deps -m pkg-deps.mk

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>