Set WType even for wireless ports without network config
nim, zedrouter and other microservices depend on the WType port
attrbiute to recognize wireless network port. This is needed even in
cases where the wireless adapter is not really used and has no network
config attached. Otherwise we may get to situation where we try to
bridge a wireless adapter or perform some other invalid operation.

In the case wireless config is not present, we can get the physical port
type from the device model and set WType accordingly.

Signed-off-by: Milan Lenco <milan@zededa.com>