installer: Add rootdelay to grub installer config
Add rootdelay to grub installer config so it can be used by the initrd
script to wait for the installer media device. This is required when
installing EVE through a Virtual Media CD/DVD-ROM emulator (found in
some BMCs) so the initrd script can wait until the virtual storage device is
detected by kernel.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>