mkimage-iso-efi: Implement rootdelay inside the initrd script
When installing EVE on devices with Virtual Media CD/DVD-ROM emulation it can
take a few seconds to kernel to detect the installer media device. This is an
issue when booting from an ISO because GRUB will boot kernel + initrd, but when
initrd script tries to find the installer media's filesystem, the Virtual
CD/DVD-ROM can be not yet available. This commit implements the same logic used
by rootdelay kernel parameter to wait for the rootfs device.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>