domainmgr: Handle VTs and Framebuffer drivers for VGA access
When VGA is enabled/disabled, the PCIe VGA device is assigned/unassigned
to/from the host through the vfio-pci driver mechanism. However, this might
not be enough for integrated graphics adapters. This commit adds additional
procedures that detach all framebuffer drivers and Virtual Terminals when
VGA access is disabled and attach them back when VGA is enabled.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>