[TO UPSTREAM]: Search for persist partition
Some specific devices (Qualcomm based) require special partition layout so
the persist partition will not be always 7 ids after IMGA. Actually this is
a hard coded value that should not be always considered as true. This
commit implements a small search of P3 partition in order to get rid of the
hard coded position.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>