GHA: select eden version based on PR target branch
This change updates the GitHub Actions workflows to select the
appropriate EDEN version based on the target branch of the PR. E.g.,
older LTS versions use older EDEN versions for testing.
The mapping is hardcoded in the workflow files. This way when we add a
new LTS branch, we only need to update a single file. Same applies when
we want to change the EDEN version for a given branch.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>